[ML] Move DFA MemoryUsage to stats.common pkg
This belongs in stats.common
